+++ src/runtime/runtime_result.py
@@ -36,6 +36,8 @@ class RTResult:
         self.old_should_return = False  # The old value of self.should_return()
 
         self.break_value: Value | None = None
+        self.break_label: str | None = None
+        self.continue_label: str | None = None
 
         self.break_or_continue_pos: tuple[Position, Position] | None = None
         self.return_pos: tuple[Position, Position] | None = None
@@ -52,6 +54,8 @@ class RTResult:
         self.loop_should_break = False
         self.old_should_return = False
         self.break_value = None
+        self.break_label = None
+        self.continue_label = None
 
         self.break_or_continue_pos = None
         self.return_pos = None
@@ -67,9 +71,14 @@ class RTResult:
         else:
             self.error = None
         self.function_return_value = result.function_return_value
+
         self.loop_should_continue = result.loop_should_continue
         self.loop_should_break = result.loop_should_break
+
         self.break_value = result.break_value
+        self.break_label = result.break_label
+        self.continue_label = result.continue_label
+
         self.break_or_continue_pos = result.break_or_continue_pos
         self.return_pos = result.return_pos
 
@@ -87,19 +96,22 @@ class RTResult:
         self.return_pos = (pos_start, pos_end)
         return self
 
-    def success_continue(self, pos_start: Position, pos_end: Position):
+    def success_continue(self, pos_start: Position, pos_end: Position, label: str | None = None):
         """same as self.success for self.loop_should_continue"""
         self.reset()
         self.break_or_continue_pos = (pos_start, pos_end)
         self.loop_should_continue = True
+        self.continue_label = label
         return self
 
-    def success_break(self, pos_start: Position, pos_end: Position, value_to_return: Value | None = None):
+    def success_break(self, pos_start: Position, pos_end: Position,
+                      value_to_return: Value | None = None, label: str | None = None):
         """same as self.success for self.loop_should_break"""
         self.reset()
         self.break_or_continue_pos = (pos_start, pos_end)
         self.loop_should_break = True
         self.break_value = value_to_return
+        self.break_label = label
         return self
 
     def failure(self, error: Error):
@@ -108,8 +120,14 @@ class RTResult:
         self.error = error
         return self
 
-    def should_return(self):  # if we should stop the interpretation because of an error, or a statement
-        #                               (return, break, continue)
+    def should_return(self, ignore_break_and_continue: bool = False):
+        """if we should stop the interpretation because of an error, or a statement
+           (return, break, continue)"""
+        if ignore_break_and_continue:
+            return (
+                self.error is not None or
+                self.function_return_value is not None
+            )
         return (
             self.error is not None or
             self.function_return_value is not None or
@@ -128,6 +146,8 @@ class RTResult:
             "loop_should_continue": self.loop_should_continue,
             "loop_should_break": self.loop_should_break,
             "break_value": self.break_value,
+            "break_label": self.break_label,
+            "continue_label": self.continue_label,
             "old_should_return": self.old_should_return,
             "break_or_continue_pos": self.break_or_continue_pos,
             "return_pos": self.return_pos